Vélocités + Le saut

+++ src/fr/ragejam/utils/Math.java
@@ -2,10 +2,10 @@ package fr.ragejam.utils;
 
 public class Math {
 	
-	public static double getIntegralPart(double value){
+	public static int getIntegralPart(double value){
 		double fractionalPart = value % 1;
 		double integralPart = value - fractionalPart;
-		return integralPart;
+		return (int) integralPart;
 	}
 
 }
